changed crypto/olm to use goolm (no cgo needed)
This package does not need to be compiled with cgo anymore.
I wrote a pure go implementation of libolm and changed crypto/olm
to use this new implementation.

Instead of completly replacing the libolm integration, the goolm
integration can now be used with the build tag 'goolm'.
Libolm is thus still the default olm implementation.